Runtime: Reorder some (at this point unused) code, bugref:10261
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@100561 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/posix/fileio-posix.cpp
@@ -351,22 +351,23 @@ RTDECL(int)  RTFileOpenEx(const char *pszFilename, uint64_t fOpen, PRTFILE phFil
         }
     }
 
+    /*
+     * If temporary file, delete it.
+     */
+    if (   fh >= 0
+        && (fOpen & RTFILE_O_TEMP_AUTO_DELETE))
+    {
+        /** @todo Use funlinkat/funlink or similar here when available!  Or better,
+         *        use O_TMPFILE, only that may require fallback as not supported by
+         *        all file system on linux. */
+        iErr = unlink(pszNativeFilename);
+        Assert(iErr == 0);
+    }
+
     rtPathFreeNative(pszNativeFilename, pszFilename);
     if (fh >= 0)
     {
         iErr = 0;
-
-        /*
-         * If temporary file, delete it.
-         */
-        if (fOpen & RTFILE_O_TEMP_AUTO_DELETE)
-        {
-            /** @todo Use funlinkat/funlink or similar here when available!  Or better,
-             *        use O_TMPFILE, only that may require fallback as not supported by
-             *        all file system on linux. */
-            iErr = unlink(pszNativeFilename);
-            Assert(iErr == 0);
-        }
 
         /*
          * Mark the file handle close on exec, unless inherit is specified.
